requirements.txt
matplotlib~=3.3.2
numpy~=1.18.5
scikit-learn~=0.23.2
tensorflow~=2.3.1
tensorflow-gpu~=2.3.1
pandas~=1.1.2
requests~=2.24.0
Pillow~=7.2.0
opencv-python~=4.4.0.44
setuptools~=50.3.0
six~=1.15.0
scikit-image~=0.17.2
Keras~=2.4.3
emnist
progressbar2